% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intensity_ops.R
\name{get_intensities}
\alias{get_intensities}
\title{Get intensiities from data that has them in columns (wide format).}
\usage{
get_intensities(D, I_col_pattern)
}
\arguments{
\item{D}{data.table with intensities and species to extract}

\item{I_col_pattern}{Pattern to match intensity columns.}
}
\value{
list with intensities and description of the experimental design
}
\description{
Get intensiities from data that has them in columns (wide format).
}
