% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intensity_ops.R
\name{get_ratios_of_medians}
\alias{get_ratios_of_medians}
\title{Get species tags from a species column.}
\usage{
get_ratios_of_medians(I, design, species, sampleComposition)
}
\arguments{
\item{I}{data.table with intensities in wide format.}

\item{design}{experimental design for columns in I.}

\item{sampleComposition}{composition of spiked in samples}
}
\value{
list with median intensities, the same filter for species in sampleComposition, and sampleCopmposition with ratios.
}
\description{
Get species tags from a species column.
}
