% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variation_plot.R
\name{get_smoothed_data}
\alias{get_smoothed_data}
\title{Get smoothed distances to reference and estimates of standard deviation.}
\usage{
get_smoothed_data(x, runs, ids, min_observed_cnt = 2, window_size = 11,
  mad_cnt = 2, spline.df = 30, grid.points.cnt = 100)
}
\arguments{
\item{x}{Array of values to smooth.}

\item{runs}{Each entries indicate the run at which the corresponding entry of x was measured.}

\item{ids}{Classifies entries of x as being the measurement of the same entity, e.g. a peptide.}

\item{min_observed_cnt}{The minimal number of runs the measurements were recorded.}

\item{window_size}{An integer describing how many points are considered in the calculation of rolling medians and MADs.}

\item{mad_cnt}{Numeric value: how many MAD from the median should be calculated for the ribbons. Non positive values result in the absence of a ribbon.}

\item{spline.df}{Degrees of freedom for the spline extrapolating the rolling medians and rolling MADs.}

\item{grid.points.cnt}{Number of points in the grid extrapolating the rolling medians and rolling MADS.}
}
\value{
A data.table with denoised and smoothed estimates of distances to the median reference run.
}
\description{
Get smoothed distances to reference and estimates of standard deviation.
}
