% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intensity_ops.R
\name{get_species}
\alias{get_species}
\title{Get species tags from a species column.}
\usage{
get_species(species_col, species_pattern = ".*_(.*)")
}
\arguments{
\item{species_col}{Vector with species to extract.}

\item{species_pattern}{Pattern to distinguish species in species column.}
}
\value{
character vector with species
}
\description{
Get species tags from a species column.
}
