% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hye.R
\name{hye_plots}
\alias{hye_plots}
\title{Plot relative ratios of the mixed proteomes.}
\usage{
hye_plots(path, organisms,
  condition_replacement = "intensity in HYE110_", add_trend = FALSE)
}
\arguments{
\item{path}{Path to the human-yeast-ecoli report.}

\item{organisms}{A data.frame with columns 'name' (of the species), 'A' (relative ratios of proteome in sample A), and 'B' (relative ratios of proteome in sample B).}

\item{condition_replacement}{Part of the column name with intensities in the original report, that must be replaced by an empty string.}

\item{add_trend}{Add in lines describing the relationship between median intensity and the ratios of median intensities.}
}
\value{
A list with two ggplot objects.
}
\description{
Plot relative ratios of the mixed proteomes.
}
