% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_sheet_ops.R
\name{isoquant_project_csv}
\alias{isoquant_project_csv}
\title{Prepare project csv with data for ISOQuant.}
\usage{
isoquant_project_csv(
  acquired_name,
  sample_description,
  folderNo,
  path_prefix = "Y:\\\\RES",
  windows_sep = T
)
}
\arguments{
\item{acquired_name}{Sample Sheet 'File name'.}

\item{sample_description}{Sample Sheet 'File Text'}

\item{folderNo}{E.g. "c('2019-050', '2019-050', '2019-050', ...)"}
}
\value{
data.table needed for IsoQuant projects.
}
\description{
Prepare project csv with data for ISOQuant.
}
