% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{lread}
\alias{lread}
\title{Read a objects from paths to a named list.}
\usage{
lread(paths, reader, ...)
}
\arguments{
\item{paths}{Array of paths.}

\item{reader}{Function used to read objects stored under paths.}

\item{...}{Further arguments to the reader.}
}
\value{
A list of opened objects.
}
\description{
Reads objects from paths and makes a list with opened objects named after the paths.
}
