% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{peptide_quantification_report}
\alias{peptide_quantification_report}
\title{Parse a ISOQuant csv peptide quantification report.}
\usage{
peptide_quantification_report(path, col_pattern)
}
\arguments{
\item{path}{Path to the report.}

\item{col_pattern}{A pattern that will select columns with intensities.}
}
\value{
Standard long format data frame object.
}
\description{
Prepare the data in the long format, so that it is easier to combine it.
}
