% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variation_plot.R
\name{plot_dist_to_reference}
\alias{plot_dist_to_reference}
\title{Plot distances to the reference run.}
\usage{
plot_dist_to_reference(S)
}
\arguments{
\item{S}{A data.table (or data.frame) with columns 'run', 'x' (the reference), and 'mid'. Additionally, can contain columns 'top' and 'bot' for ribbons. As produced by the 'get_smoothed_data' function.}
}
\value{
A ggplot object.
}
\description{
Plot distances to the reference run.
}
