% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hye.R
\name{plot_proteome_mix}
\alias{plot_proteome_mix}
\title{Plot proteome data.}
\usage{
plot_proteome_mix(X, ...)
}
\arguments{
\item{X}{Data.frame/table with columns 'species', 'A', 'B' (and potentially others).}

\item{...}{Other parameters to plot_proteome_mix2 (like bins) that don't appear in X already.}
}
\value{
List of three different ggplots.
}
\description{
Plot proteome data.
}
