% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixed_proteomes.R
\name{plot_ratios}
\alias{plot_ratios}
\title{Get species tags from a species column.}
\usage{
plot_ratios(clean_meds, sampleComposition)
}
\arguments{
\item{sampleComposition}{composition of spiked in samples}

\item{I}{data.table with intensities in wide format.}

\item{design}{experimental design for columns in I.}
}
\value{
list with the main plot, scatterplot, and boxplots.
}
\description{
Get species tags from a species column.
}
