% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixed_proteomes.R
\name{plot_scatter}
\alias{plot_scatter}
\title{Get species tags from a species column.}
\usage{
plot_scatter(meds, sampleComposition)
}
\arguments{
\item{meds}{Median intensities to plot.}

\item{sampleComposition}{composition of spiked in samples (with column ratios)}
}
\value{
a ggplot showing the ratios of intensities in two different conditions
}
\description{
This works on the output of 'get_ratios_of_medians'.
}
