% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hye.R
\name{preprocess_peptides_4_intensity_plots}
\alias{preprocess_peptides_4_intensity_plots}
\title{Prepare peptide data for the intensity comparison plots.}
\usage{
preprocess_peptides_4_intensity_plots(D, entry_species_sep = "_",
  cond_run_sep = " ")
}
\arguments{
\item{D}{Peptide report (columns 'modifier', 'sequence', 'entry', 'intensity' required).}

\item{entry_species_sep}{What sign separates the protein entries from the species description in the 'entry' column?}

\item{cond_run_sep}{What sign separates the condition from the run label in the 'cond' column?}
}
\value{
Data for proteome intensity plots.
}
\description{
Prepare peptide data for the intensity comparison plots.
}
