% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixed_proteomes.R
\name{preprocess_proteins_4_intensity_plots}
\alias{preprocess_proteins_4_intensity_plots}
\title{Prepare protein data for the intensity comparison plots.}
\usage{
preprocess_proteins_4_intensity_plots(
  D,
  entry_species_sep = "_",
  cond_run_sep = " "
)
}
\arguments{
\item{D}{Protein report (columns 'entry', 'cond', 'intensity' required).}

\item{entry_species_sep}{What sign separates the protein entries from the species description in the 'entry' column?}

\item{cond_run_sep}{What sign separates the condition from the run label in the 'cond' column?}
}
\value{
Data for proteome intensity plots.
}
\description{
Prepare protein data for the intensity comparison plots.
}
