% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{protein_report}
\alias{protein_report}
\title{Parse a full excel ISOQuant protein report.}
\usage{
protein_report(path, col_pattern, sheet = "TOP3 quantification",
  sample_no2bio_rep = NA)
}
\arguments{
\item{path}{Path to the protein report.}

\item{col_pattern}{A pattern that will select columns with intensities.}

\item{sheet}{Which excell sheet should be imported?}

\item{sample_no2bio_rep}{An additional data.frame mapping sample numbers to biological replicates.}
}
\value{
Standard long format data frame object.
}
\description{
Prepare the data in the long format, so that it is easier to combine it.
}
