% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_isoquant}
\alias{read_isoquant}
\title{Read an IsoQuant report.}
\usage{
read_isoquant(report = "protein", ...)
}
\arguments{
\item{report}{Which report is being read? ["protein", "peptide", "simple_protein"]}

\item{...}{Parameters to the called function.}
}
\value{
Wide or long data.table
}
\description{
Read an IsoQuant report.
}
