% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_configs.R
\name{read_isoquant_config}
\alias{read_isoquant_config}
\title{Read ISOQuant configuration.}
\usage{
read_isoquant_config(path, ...)
}
\arguments{
\item{path}{Path to the config file (output of ISOQuant config exporter) or a protein quantification report.}

\item{Additional}{parameters to the specific reader.}
}
\value{
data.table with config pairs: parameter-value
}
\description{
Read in the parameters of a project.
}
