% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_configs.R
\name{read_isoquant_config_from_ini}
\alias{read_isoquant_config_from_ini}
\title{Read ISOQuant ini configuration.}
\usage{
read_isoquant_config_from_ini(path)
}
\arguments{
\item{path}{Path to the config file (output of ISOQuant config exporter).}
}
\value{
data.table with config pairs: parameter-value
}
\description{
Read a separate .ini file containing ISOQuant configuration.
}
