% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_configs.R
\name{read_isoquant_config_from_report}
\alias{read_isoquant_config_from_report}
\title{Open ISOQuant config.}
\usage{
read_isoquant_config_from_report(path)
}
\value{
data.table
}
\description{
Open ISOQuant config from an excel protein quantification report.
}
