% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_configs.R
\name{read_isoquant_configs}
\alias{read_isoquant_configs}
\title{Read ISOQuant configurations.}
\usage{
read_isoquant_configs(x)
}
\arguments{
\item{paths}{Paths to the config file (output of ISOQuant config exporter) or a protein quantification report. If one path is submitted, the output will be a single data.table instead of a list of data.tables.}

\item{Additional}{parameters to the specific reader.}
}
\value{
list of data.tables or a single data.table with config pairs: parameter-value
}
\description{
Read in the parameters from either ini files or protein quantification reports.
}
