% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wide2long.R
\name{read_isoquant_protein_reports}
\alias{read_isoquant_protein_reports}
\title{Read and combine multiple reports.}
\usage{
read_isoquant_protein_reports(reports, ...)
}
\arguments{
\item{...}{Further parameters to read_isoquant_protein_report.}

\item{report}{Character vector with paths to the read in.}
}
\value{
A data.table with combined results.
}
\description{
Read in, transform from wide to long format, and bind the results togeter.
Add in the 'file' column to represent the original report.
}
