% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_report}
\alias{read_report}
\title{Read a file with intensities (peptide/protein/whatever).}
\usage{
read_report(
  path,
  I_col_pattern = "",
  sheet = "TOP3 quantification",
  drop_na_columns = T
)
}
\arguments{
\item{path}{Path to the report.}

\item{...}{Optional parameters for the wide2long function.}
}
\value{
Wide or long data.table
}
\description{
Optionally, collect the intensity columns into one column (long format).
}
