% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_wide_report}
\alias{read_wide_report}
\title{Read a file with intensities (peptide/protein/whatever) in columns (wide format).}
\usage{
read_wide_report(path, drop_na_columns = T, ...)
}
\arguments{
\item{path}{Path to the report.}

\item{drop_na_columns}{Drop empty columns.}

\item{...}{Optional parameters for either data.table::fread or readxl::read_excel}
}
\value{
Wide data.table.
}
\description{
All empty cells are replaced with NAs.
Optionally, columns entirely filled with NAs are droped.
}
