% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simple_protein_report}
\alias{simple_protein_report}
\title{An example of a simple protein report prepared with IsoQuant.}
\format{A data.table with 3118 rows and 9 variables:
\describe{
  \item{entry}{The protein PDB entry}
  \item{A 1}{Intensities in condition A, first replicate}
  \item{A 2}{Intensities in condition A, second replicate}
  ...
}}
\usage{
simple_protein_report
}
\description{
The dataset contains protein intensities gathered under two conditions (A and B).
Each condition contains a hybrid of proteome samples prepared in known ratios.
In each condition, four technical replicates of measurements were performed.
}
\keyword{datasets}
