% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wide2long.R
\name{wide2long}
\alias{wide2long}
\title{Reshape a wide report into a long format.}
\usage{
wide2long(wide_report, I_col_pattern, I_col_pattern_group_names = NA)
}
\arguments{
\item{wide_report}{A report in a wide format.}

\item{I_col_pattern}{A pattern selecting columns with intensities; consult the `stringr` package.}

\item{I_col_pattern_group_names}{Names of the groups defined in the intensity pattern. If left NA, group names will be automatically generated.}
}
\value{
A report in a long format.
}
\description{
Data in wide format has one peptide/protein per row, and multple columns with intensities.
Data in long format has one intensity per row and more than one row corresponds to the same molecule.
The long format is convenient for merging multiple reports and for plotting with ggplot2.
}
\examples{
# Columns 'intensity A 1', 'intensity A 2', 'intensity B 1', and 'intensity A 2'.
# wide2long(report, 'intensity (.) (.)', c('tech_repl', 'bio_repl'))
# Columns 'I111', 'I112','I121', 'I122', 'I311', 'I312', 'I321'
# wide2long(report, 'I(.)(.)(.)')
}
