% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_isoquant_peptide_report}
\alias{read_isoquant_peptide_report}
\title{Read an ISOQuant csv peptide report.}
\usage{
read_isoquant_peptide_report(path, ...)
}
\arguments{
\item{path}{Path to the report.}

\item{...}{Optional parameters for the wide2long function.}
}
\value{
Wide or long data.table
}
\description{
Read an ISOQuant csv peptide report.
}
