% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_isoquant_peptide_report}
\alias{read_isoquant_peptide_report}
\title{Read an ISOQuant csv peptide report.}
\usage{
read_isoquant_peptide_report(path, long_df = FALSE,
  intensity_pattern = "intensity in")
}
\arguments{
\item{path}{Path to the report.}

\item{long_df}{Should the output be in the long format? [default = FALSE]}

\item{intensity_pattern}{A pattern that will select columns with intensities [default="intensity in"]}
}
\value{
Wide or long data.table
}
\description{
Read an ISOQuant csv peptide report.
}
