% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_isoquant_protein_report}
\alias{read_isoquant_protein_report}
\title{Read an ISOQuant protein report.}
\usage{
read_isoquant_protein_report(path, sheet = "TOP3 quantification", ...)
}
\arguments{
\item{path}{Path to the protein report.}

\item{sheet}{Which excell sheet should be imported [default="TOP3 quantification"]}

\item{...}{Optional parameters for the wide2long function.}
}
\value{
Wide or long data.table
}
\description{
Read an ISOQuant protein report.
}
