% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_isoquant_protein_report}
\alias{read_isoquant_protein_report}
\title{Read an ISOQuant protein report.}
\usage{
read_isoquant_protein_report(path, long_df = FALSE,
  intensity_pattern = "[A|B][:space:][:digit:]",
  sheet = "TOP3 quantification")
}
\arguments{
\item{path}{Path to the protein report.}

\item{long_df}{Should the output be in the long format? [default = FALSE]}

\item{intensity_pattern}{A pattern that will select columns with intensities [default = '[A|B][:space:][:digit:]']}

\item{sheet}{Which excell sheet should be imported [default="TOP3 quantification"]}

\item{...}{Other parameters that are meaningless.}
}
\value{
Wide or long data.table
}
\description{
Read an ISOQuant protein report.
}
