% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_isoquant_simple_protein_report}
\alias{read_isoquant_simple_protein_report}
\title{Read a simple protein ISOQuant report.}
\usage{
read_isoquant_simple_protein_report(path, long_df = FALSE,
  prot_cols = "entry")
}
\arguments{
\item{path}{Path to the report.}

\item{long_df}{Should the output be in the long format? [default = FALSE]}

\item{prot_cols}{Columns identifying proteins.}
}
\value{
Wide or long data.table
}
\description{
Read a simple protein ISOQuant report.
}
