% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{simple_protein_report}
\alias{simple_protein_report}
\title{Parse a simple ISOQuant csv protein quantification report.}
\usage{
simple_protein_report(path)
}
\arguments{
\item{path}{Path to the report.}
}
\value{
Standard long format data frame object.
}
\description{
Prepare the data in the long format, so that it is easier to combine it.
}
